updated the strategies to use a inheritance, to reduce the amount of repeated logic as we can just call super()... instead of redifining things, or putting them in helper functions.
