+++ fuzzball/modules/helper.py
@@ -16,10 +16,3 @@ def get_arguments():
         print(f'[*] binary file: {os.path.basename(binary)}. sample input: {os.path.basename(sample)}')
 
     return binary, sample
-
-def get_random_string(length):
-    return ''.join(random.choice(string.ascii_lowercase + string.ascii_uppercase) for _ in range(length))
-
-# TODO: fix
-def get_random_format_string(size):
-    return ''.join(random.choice(['%x', '%c', '%d', '%p', '%s']) for _ in range(size))
