--- fuzzball/strategies/common.py
@@ -1,4 +1,0 @@
-class CommonStrategy:
-    def generate_input(self):
-        yield b''
-        yield b'1'
